requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.0
spacy==3.1.2
